refactor: complete Phase 2 - streaming operations modularization and TypeScript fixes
Phase 2.1: Domain-driven streaming composables restructuring
- Split useStreamingOperations.ts (432 lines → 75 lines, 83% reduction)
- Created 3 domain-specific composables:
  * useBasicStreaming.ts (149 lines): Core streaming functionality
  * useStreamingWithConfig.ts (98 lines): Configuration-based streaming
  * useStreamingWithTools.ts (214 lines): MCP tools integration streaming
- Implemented composable aggregation pattern maintaining backward compatibility
- Extracted shared types and interfaces to streaming/index.ts

TypeScript error elimination (100% resolution):
- Fixed hierarchical LLM request structure (query.conversation_history)
- Updated streaming callback interfaces (onToolCall, onToolResult)
- Resolved Vue component type definitions (activeToolExecutions)
- Fixed Repository type compatibility issues
- Updated test files for new API structure
- Corrected LLM service return types and capabilities

API service improvements:
- Enhanced type safety with components['schemas'] pattern
- Added fallback type definitions for missing schemas
- Updated streaming clients to use proper type hierarchy
- Fixed tools service configuration structure

Code quality improvements:
- Eliminated all remaining any types where possible
- Added proper TypeScript generics and constraints
- Updated test mock structures for new architecture
- Maintained 100% backward compatibility

Result: Zero TypeScript errors, improved maintainability, enhanced modularity
Ready for Phase 3 component extraction and further architectural improvements.
