+++ src/services/api/api.service.ts
@@ -5,21 +5,63 @@
  * 後方互換性を保ちつつ、内部的にはドメイン別クライアントを使用
  */
 import { ApiClientFactory } from './api-client.factory'
-import type {
-  DocumentResponse,
-  RepositoryStructureResponse,
-  RepositoryResponse,
-  RepositoryCreate,
-  RepositoryUpdate,
-  SearchQuery,
-  SearchResponse,
-  LLMQueryRequest,
-  LLMResponse,
-  LLMStreamingRequest,
-  StreamingLLMResponse,
-  MCPToolsResponse,
-  MCPToolInfo
-} from './types'
+import type { components } from './types.auto'
+
+type DocumentResponse = components['schemas']['DocumentResponse']
+type RepositoryStructureResponse = components['schemas']['RepositoryStructureResponse']
+type LLMQueryRequest = components['schemas']['LLMQueryRequest']
+type LLMResponse = components['schemas']['LLMResponse']
+type MCPToolsResponse = components['schemas']['MCPToolsResponse']
+type MCPToolInfo = components['schemas']['MCPToolInfo']
+
+// Fallback types for missing schemas
+interface RepositoryResponse {
+  id: number
+  name: string
+  owner: string
+  service: string
+  created_at: string
+  updated_at: string
+}
+
+interface RepositoryCreate {
+  name: string
+  owner: string
+  service: string
+  description?: string
+}
+
+interface RepositoryUpdate {
+  name?: string
+  description?: string
+}
+
+interface SearchQuery {
+  query: string
+  limit: number
+  offset: number
+}
+
+interface SearchResponse {
+  results: any[]
+  total: number
+}
+
+interface LLMStreamingRequest {
+  prompt: string
+  provider: string
+  model?: string
+  conversation_history?: components['schemas']['MessageItem'][]
+  stream?: boolean
+}
+
+interface StreamingLLMResponse {
+  data: {
+    content?: string
+    error?: string
+    [key: string]: any
+  }
+}
 
 export class ApiService {
   private factory: ApiClientFactory
